# Use `RSTRING_PTR()` to retrieve C string pointer