--- a/ext/oj/parse.c
+++ b/ext/oj/parse.c
@@ -969,7 +969,7 @@ static void oj_pi_set_input_str(ParseInfo pi, volatile VALUE *inputp) {
     if (rb_utf8_encoding() != enc) {
         *inputp = rb_str_conv_enc(*inputp, enc, rb_utf8_encoding());
     }
-    pi->json = rb_string_value_ptr((VALUE *)inputp);
+    pi->json = RSTRING_PTR(*inputp);
     pi->end  = pi->json + RSTRING_LEN(*inputp);
 }